requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
twilio
httpx
python-multipart
jinja2
